Reorganize the new crate a bit
